Ignore the "In Progress" folder
This folder contains the writeups for challenges I have not yet solved. The writeups are created as I solve the problem to help me visualize better the information I have gathered about the problem. The writeups will be moved to another folder and uploaded as soon as they are completed.