Fixed phrasing errors
I read through my writeup, and I felt that my phrasing was weird. So I made amendments, especially to the part where I state "We have attached pictures from 2 separate runs to let you see what the password looks like". At the time, we did not know that it was the password YET.