Fixed images and hyperlink
Some images were too large (in dimension), so I cropped out the unimportant parts. That way, readers would not be distracted either. Also fixed a hyperlink error to the file iot_ctf2018_serially_fun_flag.jpg. I forgot to add the path traversal to the readme file so it was referencing a non-existant file.