Editted writeup - Included screenshot of APIs
We did not take screenshots of the API when we did the challenge, so the writeups was not quite complete. However, the creator of the challenge informed us that the documentation is also available online in a very similar (if not the same) format. We realised it was the SAME type of documentation, so we took it. The screenshots are now included in the writeups for the 2 challenges involving the smart TV.