Added recently discovered fact
I was playing around with the board and I realised that sometimes the password is not all the extra characters printed. One had to try a few times to get it right.
I accidentally omitted the fact that sometimes the characters printed are illegible too. It could be important to others trying out the challenge at home.